refactor(rocksdb): remove rocksdb, grpc service, and related commands
