+++ Cargo.toml
@@ -2,11 +2,12 @@
 name = "thegarii"
 authors = ['info@chainsafe.io']
 description = 'thegarii firehose service'
-version = "0.0.3"
+version = "0.0.4"
 license = "GPL-3.0"
 homepage = 'https://github.com/ChainSafe/the-garii'
 repository = 'https://github.com/ChainSafe/the-garii'
 edition = "2021"
+readme = "./README.md"
 
 [[bin]]
 name = "thegarii"
@@ -24,23 +25,16 @@ prost = "0.9"
 prost-types = "0.9"
 rand = "0.8.5"
 reqwest = { version = "0.11.9", features = ["gzip", "json"] }
-rocksdb = { version = "0.18.0", optional = true }
 serde = { version = "^1", features = ["derive"] }
 serde_json = "1.0.79"
 structopt = "0.3.26"
 thiserror = "1.0.30"
 tonic = { version = "0.6.2", optional = true }
-tonic-web = { version = "0.2.0", optional = true }
 tokio = { version = "1", features = ["full"] }
 
 [features]
-default = [ "firehose" ]
-firehose = [ ]
-full = [
-     "rocksdb",
-     "tonic",
-     "tonic-web"
-]
+default = [ ]
+stream = [ "tonic" ]
 
 [build-dependencies]
 tonic-build = "0.6.2"
