+++ src/cmd/get.rs
@@ -1,7 +1,6 @@
 // Copyright 2021 ChainSafe Systems
 // SPDX-License-Identifier: LGPL-3.0-only
-use crate::{cmd::CommandT, Client, Env, Result, Storage};
-use async_trait::async_trait;
+use crate::{Client, Result};
 use structopt::StructOpt;
 
 #[derive(StructOpt, Debug)]
@@ -10,18 +9,10 @@ pub struct Get {
     pub height: u64,
 }
 
-#[async_trait]
-impl CommandT for Get {
-    async fn exec(&self, env: Env) -> Result<()> {
-        let storage = Storage::read_only(&env.db_path)?;
-
-        let block = if let Ok(block) = storage.get(self.height) {
-            block
-        } else {
-            log::warn!("block not exists, fetching from endpoints...");
-            let client = Client::from_env()?;
-            client.get_firehose_block_by_height(self.height).await?
-        };
+impl Get {
+    pub async fn exec(&self) -> Result<()> {
+        let client = Client::from_env()?;
+        let block = client.get_firehose_block_by_height(self.height).await?;
 
         println!("{}", serde_json::to_string_pretty(&block)?);
         Ok(())
