+++ src/cmd/mod.rs
@@ -1,45 +1,26 @@
 // Copyright 2021 ChainSafe Systems
 // SPDX-License-Identifier: LGPL-3.0-only
-#![cfg(feature = "full")]
 
 //! thegarii commands
 use crate::{Env, EnvArguments, Result};
-use async_trait::async_trait;
 use structopt::StructOpt;
 
-mod backup;
 mod console;
 mod get;
 mod poll;
-mod restore;
-mod start;
+
+#[cfg(feature = "stream")]
 mod stream;
-mod syncing;
 
 #[derive(StructOpt, Debug)]
 pub enum Command {
-    /// Backup blocks to path
-    Backup(backup::Backup),
-    /// Polling blocks and write to stdout
-    Console(console::Console),
     /// Get a block from database or fetch it
     Get(get::Get),
     /// Dry-run random polling with time estimate
     Poll(poll::Poll),
-    /// Restore blocks from path
-    Restore(restore::Restore),
-    /// Start thegarii service
-    Start(start::Start),
     /// Stream blocks from gRPC service
+    #[cfg(feature = "stream")]
     Stream(stream::Stream),
-    /// Show the syncing status
-    Syncing(syncing::Syncing),
-}
-
-/// Command trait
-#[async_trait]
-pub trait CommandT {
-    async fn exec(&self, env: Env) -> Result<()>;
 }
 
 #[derive(StructOpt, Debug)]
@@ -54,7 +35,7 @@ pub struct Opt {
 
     /// commands
     #[structopt(subcommand)]
-    pub command: Command,
+    pub command: Option<Command>,
 }
 
 impl Opt {
@@ -71,15 +52,15 @@ impl Opt {
         }
 
         let env = Env::from_args(opt.env)?;
-        match opt.command {
-            Command::Backup(backup) => backup.exec(env).await?,
-            Command::Console(console) => console.exec(env).await?,
-            Command::Get(get) => get.exec(env).await?,
-            Command::Poll(poll) => poll.exec(env).await?,
-            Command::Restore(restore) => restore.exec(env).await?,
-            Command::Start(start) => start.exec(env).await?,
-            Command::Stream(stream) => stream.exec(env).await?,
-            Command::Syncing(syncing) => syncing.exec(env).await?,
+        if let Some(cmd) = opt.command {
+            match cmd {
+                Command::Get(get) => get.exec().await?,
+                Command::Poll(poll) => poll.exec(env).await?,
+                #[cfg(feature = "stream")]
+                Command::Stream(stream) => stream.exec().await?,
+            }
+        } else {
+            console::Console::new(env)?.exec().await?;
         }
 
         Ok(())
