+++ src/lib.rs
@@ -5,32 +5,16 @@
 //!
 //! this library fetches blocks from Arweave and generates firehose blocks for the Graph
 pub mod client;
-mod console;
+pub mod cmd;
 mod encoding;
 pub mod env;
 pub mod pb;
 pub mod result;
 pub mod types;
 
-#[cfg(feature = "full")]
-pub mod cmd;
-#[cfg(feature = "full")]
-pub mod service;
-#[cfg(feature = "full")]
-mod storage;
-
-#[cfg(feature = "full")]
-pub use storage::Storage;
-
-#[cfg(not(feature = "firehose"))]
-pub use cmd::Opt;
-
-#[cfg(feature = "firehose")]
-pub use console::cmd::Opt;
-
 pub use self::{
     client::Client,
-    console::Console,
+    cmd::Opt,
     env::{Env, EnvArguments},
     result::{Error, Result},
 };
