+++ src/pb.rs
@@ -18,7 +18,7 @@ pub mod sf {
         }
     }
 
-    #[cfg(feature = "full")]
+    #[cfg(feature = "stream")]
     pub mod firehose {
         pub mod v1 {
             include!(concat!(env!("OUT_DIR"), "/sf.firehose.v1.rs"));
@@ -28,7 +28,7 @@ pub mod sf {
 
 pub use self::sf::arweave::r#type::v1::*;
 
-#[cfg(feature = "full")]
+#[cfg(feature = "stream")]
 pub use self::sf::firehose::v1::*;
 
 /// decode string to bytes with base64url
